+++ test/mock/01-ok.js
@@ -3,12 +3,12 @@ const id = 'mock_one'
 
 module.exports = {
   async up (db, logger) {
-    logger.verbose('Inserting id %s into collection %s.', id, collection)
+    logger.debug('Inserting id %s into collection %s.', id, collection)
 
     await db.collection(collection).insertOne({ id })
   },
   async down (db, logger) {
-    logger.verbose('Removing id %s from collection %s.', id, collection)
+    logger.debug('Removing id %s from collection %s.', id, collection)
 
     await db.collection(collection).removeOne({ id })
   }